Linux: Fix case of user defined loopback
* Add loopback ifname definition for Linux devices
* Move setting of 'af' out of link transformation, also invoke in case of no links
* Abort loop in 'set_node_af' correctly
