+++ netsim/devices/linux.yml
@@ -1,6 +1,7 @@
 description: Generic Linux host
 interface_name: eth{ifindex}
 lag_interface_name: "bond{lag.ifindex}"
+loopback_interface_name: lo
 mgmt_if: eth0
 role: host
 features:
